armpilot/backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11.7"
dependencies = [
    "fastapi>=0.135.1",
    "lerobot>=0.4.4",
    "numpy>=2.4.3",
    "openai>=2.28.0",
    "pydantic>=2.12.5",
    "python-dotenv>=1.2.2",
    "scservo-sdk",
    "tavily-python>=0.7.23",
    "uvicorn[standard]>=0.41.0",
    "websockets>=16.0",
]
